+++ lib/brazil.dart
@@ -2,4 +2,5 @@
 library brazil;
 
 export 'src/brazil/cep_field.dart';
+export 'src/brazil/cnpj_field.dart';
 export 'src/brazil/cpf_field.dart';